+++ tests/test_sign_cmd.py
@@ -323,3 +323,42 @@ def test_sign_legacy_tokens(firmware, backend, navigator, test_name):
     response = client.get_async_response().data
     _, sig, _ = unpack_sign_tx_response(response)
     assert check_signature_validity(public_key, sig, transaction)
+
+# # In this test we send to the device a transaction to sign and validate it on screen
+# # The transaction is Fungible Asset Tokens and should be Clear Signed
+def test_sign_fa_tx(firmware, backend, navigator, test_name):
+    # Use the app interface instead of raw interface
+    client = AptosCommandSender(backend)
+    # The path used for this entire test
+    path: str = "m/44'/637'/1'/0'/0'"
+
+    # First we need to get the public key of the device in order to build the transaction
+    rapdu = client.get_public_key(path=path)
+    _, public_key, _, _ = unpack_get_public_key_response(rapdu.data)
+
+    # Create the transaction that will be sent to the device for signing
+    transaction = bytes.fromhex("b5e97db07fa0bd0e5598aa3643a9bc6f6693bddc1a9fec9e674a461eaa00b1938f13f355f3af444bd356adeaaaf01235a7817d6a4417f5c9fa3d74a68f7b7afd0000000000000000020000000000000000000000000000000000000000000000000000000000000001167072696d6172795f66756e6769626c655f73746f7265087472616e73666572010700000000000000000000000000000000000000000000000000000000000000010e66756e6769626c655f6173736574084d65746164617461000320357b0b74bc833e95a115ad22604854d6b0fca151cecd94111770e5d6ffc9dc2b207be51d04d3a482fa056bc094bc5eadad005aaf823a95269410f08730f0d03cb40840420f000000000009000000000000006400000000000000000000000000000001")
+    
+    # Send the sign device instruction.
+    # As it requires on-screen validation, the function is asynchronous.
+    # It will yield the result when the navigation is done
+    with client.sign_tx(path=path, transaction=transaction):
+        # Validate the on-screen request by performing the navigation appropriate for this device
+        if firmware.device.startswith("nano"):
+            navigator.navigate_until_text_and_compare(NavInsID.RIGHT_CLICK,
+                                                      [NavInsID.BOTH_CLICK],
+                                                      "Approve",
+                                                      ROOT_SCREENSHOT_PATH,
+                                                      test_name)
+        else:
+            navigator.navigate_until_text_and_compare(NavInsID.USE_CASE_VIEW_DETAILS_NEXT,
+                                                      [NavInsID.USE_CASE_REVIEW_CONFIRM,
+                                                       NavInsID.USE_CASE_STATUS_DISMISS],
+                                                      "Hold to sign",
+                                                      ROOT_SCREENSHOT_PATH,
+                                                      test_name)
+
+    # The device as yielded the result, parse it and ensure that the signature is correct
+    response = client.get_async_response().data
+    _, sig, _ = unpack_sign_tx_response(response)
+    assert check_signature_validity(public_key, sig, transaction)
